requirements.txt
discord.py
jinja2
discord.py
jinja2
Pillow
beautifulsoup4
requests
python-dotenv
